Migrate add/remove maps window to other cup views
